+++ src/main/java/com/twentyone/steachserver/domain/statistic/repository/LectureStatisticMongoRepository.java
@@ -1,6 +1,6 @@
 package com.twentyone.steachserver.domain.statistic.repository;
 
-import com.twentyone.steachserver.domain.statistic.model.LectureStatisticsByAllStudent;
+import com.twentyone.steachserver.domain.statistic.model.mongoDB.LectureStatisticsByAllStudent;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
