+++ src/main/java/com/twentyone/steachserver/domain/statistic/service/StatisticServiceImpl.java
@@ -6,8 +6,8 @@ import com.twentyone.steachserver.domain.lecture.model.Lecture;
 import com.twentyone.steachserver.domain.member.model.Student;
 import com.twentyone.steachserver.domain.statistic.dto.radarChartStatisticDto;
 import com.twentyone.steachserver.domain.statistic.dto.GPTDataRequestDto;
-import com.twentyone.steachserver.domain.statistic.model.GPTDataByLecture;
-import com.twentyone.steachserver.domain.statistic.model.LectureStatisticsByAllStudent;
+import com.twentyone.steachserver.domain.statistic.model.mongoDB.GPTDataByLecture;
+import com.twentyone.steachserver.domain.statistic.model.mongoDB.LectureStatisticsByAllStudent;
 import com.twentyone.steachserver.domain.statistic.dto.temp.LectureStatisticsByStudentDto;
 import com.twentyone.steachserver.domain.statistic.repository.GPTDataByLectureMongoRepository;
 import com.twentyone.steachserver.domain.statistic.repository.LectureStatisticMongoRepository;
